Make the export easier to grok

+++ src/Command/ExportCommand.php
@@ -8,18 +8,24 @@ use Meanbee\Magedbm2\Shell\Command\Gzip;
 use Meanbee\Magedbm2\Shell\Command\Mysqldump;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
 class ExportCommand extends BaseCommand
 {
     const NAME            = 'export';
     const ARG_OUTPUT_FILE = 'output-file';
+    const OPT_NO_COMPRESS = 'no-compress';
 
     /**
      * @var ConfigInterface
      */
     private $config;
 
+    /**
+     * @var Anonymiser
+     */
+    private $anonymiser;
 
     /**
      * @param ConfigInterface $config
@@ -29,6 +35,7 @@ class ExportCommand extends BaseCommand
     {
         parent::__construct(self::NAME);
         $this->config = $config;
+        $this->anonymiser = new Anonymiser();
     }
 
     /**
@@ -43,82 +50,28 @@ class ExportCommand extends BaseCommand
             return $parentExitCode;
         }
 
-        $anonymiser = new Anonymiser();
-        $anonymiser->setLogger($this->getLogger());
+        $this->anonymiser->setLogger($this->getLogger());
 
-        $config = $this->config->get('anonymizer');
+        $tablesToExport = $this->configureAnonymiser();
 
-        $tables = [
-            'eav_entity_type',
-            'eav_attribute'
-        ];
-
-        foreach ($config['tables'] as $tableConfig) {
-            $tableName = $tableConfig['name'];
-
-            foreach ($tableConfig['columns'] as $columnName => $formatter) {
-                $anonymiser->addColumnRule($tableName, $columnName, $formatter);
-            }
-
-            $tables[] = $tableName;
-        }
-
-        foreach ($config['eav'] as $eavConfig) {
-            $entityName = $eavConfig['entity'];
-
-            foreach ($eavConfig['attributes'] as $attributeName => $formatter) {
-                $anonymiser->addAttributeRule($entityName, $attributeName, $formatter);
-            }
-
-            foreach (['datetime', 'decimal', 'int', 'text', 'varchar'] as $type) {
-                $tables[] = sprintf('%s_entity_%s', $entityName, $type);
-            }
-        }
-
-        $tables = array_unique($tables);
-
-        $exportFile = $this->generateTemporaryFile();
-        $temporaryFileName = $this->generateTemporaryFile();
         $outputFileName = $this->input->getArgument(self::ARG_OUTPUT_FILE);
+        $temporaryFileName = $this->generateTemporaryFile();
 
-        $this->getLogger()->info('Generating XML dump from database');
+        $exportFile = $this->generateDatabaseExport($tablesToExport);
 
-        $command = (new Mysqldump())
-            ->arguments($this->getCredentialOptions())
-            ->argument('--xml')
-            ->argument($this->config->getDatabaseCredentials()->getName())
-            ->arguments($tables)
-            ->output($exportFile);
-
-        $this->getLogger()->debug($command->toString());
-
-        $process = $command->toProcess();
-        $process->start();
-        $process->wait();
-
-        $this->getLogger()->info('Dump completed');
-
-        $anonymiser->processFile($exportFile, $temporaryFileName);
+        $this->anonymiser->processFile($exportFile, $temporaryFileName);
 
         unlink($exportFile);
 
         $this->getLogger()->info('Anonymised dump at ' . $temporaryFileName);
 
-        $gzip = (new Gzip())
-            ->argument('-9')
-            ->argument('--force')
-            ->argument('--to-stdout')
-            ->argument($temporaryFileName)
-            ->output($outputFileName);
+        if ($this->shouldCompress()) {
+            $this->compressFile($temporaryFileName, $outputFileName);
 
-        $gzipProcess = $gzip->toProcess();
-
-        $this->getLogger()->debug($gzip->toString());
-
-        $gzipProcess->start();
-        $gzipProcess->wait();
-
-        unlink($temporaryFileName);
+            unlink($temporaryFileName);
+        } else {
+            rename($temporaryFileName, $outputFileName);
+        }
 
         $this->output->writeln("Exported to $outputFileName");
 
@@ -131,7 +84,8 @@ class ExportCommand extends BaseCommand
             ->setName(self::NAME)
             ->setDescription('Create an anonymised data export of sensitive tables');
 
-        $this->addArgument(self::ARG_OUTPUT_FILE, InputArgument::REQUIRED, 'Location to output the anonymised export');
+        $this->addArgument(self::ARG_OUTPUT_FILE, InputArgument::REQUIRED, 'Location to output the anonymised export.');
+        $this->addOption(self::OPT_NO_COMPRESS, 'Z', InputOption::VALUE_NONE, 'Disable gzipping of export file.');
     }
 
     /**
@@ -172,5 +126,118 @@ class ExportCommand extends BaseCommand
         }
 
         return $file;
+    }
+
+    /**
+     * @return bool
+     */
+    private function shouldCompress()
+    {
+        return ((bool) $this->input->getOption(self::OPT_NO_COMPRESS)) === false;
+    }
+
+    /**
+     * @param $tables
+     * @return string
+     * @throws \Exception
+     */
+    protected function generateDatabaseExport($tables): string
+    {
+        $exportFile = $this->generateTemporaryFile();
+
+        $this->getLogger()->info('Generating XML dump from database');
+
+        $command = (new Mysqldump())
+            ->arguments($this->getCredentialOptions())
+            ->argument('--xml')
+            ->argument($this->config->getDatabaseCredentials()->getName())
+            ->arguments($tables)
+            ->output($exportFile);
+
+        $this->getLogger()->debug($command->toString());
+
+        $process = $command->toProcess();
+        $process->start();
+        $process->wait();
+
+        if ($process->getExitCode() !== 0) {
+            throw new \RuntimeException(sprintf(
+                'Mysqldump: Process failed with code %s',
+                $process->getExitCode()
+            ));
+        }
+
+        $this->getLogger()->info('Dump completed');
+
+        if (!file_exists($exportFile)) {
+            throw new \RuntimeException('Dump file was not created');
+        }
+
+        if (!is_readable($exportFile)) {
+            throw new \RuntimeException('Created dump file was not readable');
+        }
+        return $exportFile;
+    }
+
+    /**
+     * @param $inputFile
+     * @param $outputFile
+     */
+    protected function compressFile($inputFile, $outputFile): void
+    {
+        $gzip = (new Gzip())
+            ->argument('-9')
+            ->argument('--force')
+            ->argument('--to-stdout')
+            ->argument($inputFile)
+            ->output($outputFile);
+
+        $gzipProcess = $gzip->toProcess();
+
+        $this->getLogger()->debug($gzip->toString());
+
+        $gzipProcess->start();
+        $gzipProcess->wait();
+    }
+
+    /**
+     * @return array
+     */
+    protected function configureAnonymiser(): array
+    {
+        $config = $this->config->get('anonymizer');
+
+        $tables = [
+            'eav_entity_type',
+            'eav_attribute'
+        ];
+
+        foreach ($config['tables'] as $tableConfig) {
+            $tableName = $tableConfig['name'];
+
+            foreach ($tableConfig['columns'] as $columnName => $formatter) {
+                $newTables = $this->anonymiser->addColumnRule($tableName, $columnName, $formatter);
+
+                foreach ($newTables as $table) {
+                    $tables[] = $table;
+                }
+            }
+
+            $tables[] = $tableName;
+        }
+
+        foreach ($config['eav'] as $eavConfig) {
+            $entityName = $eavConfig['entity'];
+
+            foreach ($eavConfig['attributes'] as $attributeName => $formatter) {
+                $newTables = $this->anonymiser->addAttributeRule($entityName, $attributeName, $formatter);
+
+                foreach ($newTables as $table) {
+                    $tables[] = $table;
+                }
+            }
+        }
+
+        return array_unique($tables);
     }
 }
